requirements.txt
transformers
einops
timm
accelerate
datasets
pillow
mteb[vision]
bitsandbytes
peft
faiss-gpu-cu12